Add client tests and fix listen bug (#1)
Due to the function signature the return value was being wrapped in
`AsyncGenerator` similar to a `Promise` with `async`. This PR removes
the generator signature. It also adds tests to ensure it works as
expected, previously we relied on TypeScript but that didn't catch this.
All the actual logic already has tests.